Fixing imports and moving helper functions
