+++ lib/chainsync/chainsync/base/__init__.py
@@ -5,13 +5,11 @@ from .db_schema import Base, UserMap
 from .postgres import (
     PostgresConfig,
     TableWithBlockNumber,
-    add_transactions,
     add_user_map,
     build_postgres_config,
     close_session,
     drop_table,
     get_latest_block_number_from_table,
-    get_transactions,
     get_user_map,
     initialize_engine,
     initialize_session,
